+++ app/_components/ai-panel.tsx
@@ -1,8 +1,14 @@
 "use client";
 
 import { useEffect, useMemo, useState } from "react";
-import { CheckIcon, ClipboardDocumentIcon, CloudArrowUpIcon } from "@heroicons/react/24/outline";
+
 import { BlockingOverlay } from "./blocking-overlay";
+import { AiPanelHeader } from "./ai-panel/AiPanelHeader";
+import { AiPanelInput } from "./ai-panel/AiPanelInput";
+import { AiPanelNotice } from "./ai-panel/AiPanelNotice";
+import { AiPanelResponse } from "./ai-panel/AiPanelResponse";
+import { AiResponse } from "./ai-panel/types";
+import { buildCopyText } from "./ai-panel/utils";
 
 type Props = {
   kind: "es_review" | "company_analysis" | "aptitude_analysis" | "self_analysis" | "interview_review";
@@ -18,20 +24,6 @@ type Props = {
   onSaved?: () => void;
   showOneShotNotice?: boolean;
 };
-
-type AiResponse = {
-  summary?: string;
-  bulletPoints?: string[];
-  provider?: string;
-  error?: string;
-};
-
-const sanitizeMarkdown = (text: string): string =>
-  text
-    .replace(/^#{1,6}\s*/gm, "")
-    .replace(/^\s*[-*・]\s?/gm, "・")
-    .replace(/\*\*(.*?)\*\*/g, "$1")
-    .trim();
 
 export function AiPanel({
   kind,
@@ -168,10 +160,7 @@ export function AiPanel({
 
   const handleCopy = async () => {
     if (!response) return;
-    const summary = response.summary ? sanitizeMarkdown(response.summary) : "";
-    const bullets =
-      response.bulletPoints?.length && response.bulletPoints.map((b) => `・${sanitizeMarkdown(b)}`).join("\n");
-    const text = [summary, bullets].filter(Boolean).join("\n\n");
+    const text = buildCopyText(response);
     try {
       await navigator.clipboard.writeText(text);
       setCopied(true);
@@ -187,95 +176,32 @@ export function AiPanel({
     <>
       {overlay ? <BlockingOverlay message="AI処理中です。画面を閉じずにお待ちください。" /> : null}
       <div className="relative rounded-2xl border border-slate-200 bg-white/90 p-5 text-sm text-slate-900 shadow-md dark:border-slate-700 dark:bg-slate-900/80 dark:text-slate-100">
-        {showOneShotNotice ? (
-          <div className="mb-3 rounded-xl border border-amber-200 bg-amber-50 px-3 py-2 text-[11px] font-semibold text-amber-800">
-            ※ 保存済みのAI回答は再実行できません。再度利用したい場合は運営にお問い合わせください。
-          </div>
-        ) : null}
-        <div className="flex items-center justify-between">
-          <div>
-            <p className="text-xs font-semibold text-amber-700">{title}</p>
-            {hint ? <p className="text-[11px] text-slate-500 dark:text-slate-300">{hint}</p> : null}
-            {cacheKey ? (
-              <p className="text-[11px] text-slate-400 dark:text-slate-500">
-                ※ 1回保存すると再実行はできません。入力を確認してから送信してください。
-              </p>
-            ) : null}
-          </div>
-          <span className="rounded-full border border-slate-200 bg-slate-50 px-3 py-1 text-[11px] text-slate-700 dark:border-slate-600 dark:bg-slate-800 dark:text-slate-100">
-            {response?.provider ?? "AI"}
-          </span>
-        </div>
-
-        <textarea
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          rows={8}
-          className="mt-3 w-full rounded-xl border border-slate-200 bg-white px-4 py-3 text-sm text-slate-900 outline-none focus:border-amber-300 dark:border-slate-600 dark:bg-slate-900 dark:text-slate-100"
+        <AiPanelNotice show={showOneShotNotice} />
+        <AiPanelHeader title={title} hint={hint} cacheKey={cacheKey} response={response} />
+        <AiPanelInput
+          input={input}
+          onChange={setInput}
+          wordCount={wordCount}
+          hint={hint}
+          loading={loading}
+          saved={saved}
+          saveUrl={saveUrl}
+          onRun={handleRun}
         />
-        <div className="flex items-center justify-between text-[11px] text-slate-500 dark:text-slate-300">
-          <span>単語数: {wordCount}</span>
-          {hint ? <span>{hint}</span> : null}
-        </div>
-
-        <div className="mt-3 flex justify-end">
-          <button
-            type="button"
-            onClick={handleRun}
-            disabled={loading || (saved && !!saveUrl)}
-            className="rounded-full bg-linear-to-r from-emerald-300 via-cyan-300 to-sky-300 px-4 py-2 text-xs font-semibold text-slate-950 shadow-lg shadow-emerald-400/30 transition hover:translate-y-0.5 hover:shadow-emerald-400/50 disabled:cursor-not-allowed disabled:opacity-60"
-          >
-            {loading ? "送信中..." : saved && !!saveUrl ? "保存済み" : "AIに送る"}
-          </button>
-        </div>
 
         {error ? <p className="mt-2 text-[11px] text-rose-500">Error: {error}</p> : null}
 
         {response ? (
-          <div className="mt-3 space-y-3 rounded-xl border border-slate-200 bg-slate-50 p-3 dark:border-slate-700 dark:bg-slate-800">
-            <div className="flex items-center justify-between">
-              <p className="text-xs font-semibold text-amber-700">
-                AI回答 {response?.provider === "saved" ? "（保存済み）" : ""}
-              </p>
-              <div className="flex gap-2">
-                {saveUrl && saveId && !saved && (
-                  <button
-                    type="button"
-                    onClick={handleSave}
-                    disabled={saving}
-                    className="flex items-center gap-1 rounded-full border border-emerald-500 bg-emerald-50 px-3 py-1 text-[11px] font-semibold text-emerald-700 transition hover:bg-emerald-100 disabled:cursor-not-allowed disabled:opacity-60 dark:border-emerald-600 dark:bg-emerald-900/30 dark:text-emerald-300 dark:hover:bg-emerald-900/50"
-                  >
-                    <CloudArrowUpIcon className="h-4 w-4" />
-                    <span>{saving ? "保存中..." : "保存する"}</span>
-                  </button>
-                )}
-                {saved && saveUrl && (
-                  <span className="flex items-center gap-1 rounded-full border border-emerald-500 bg-emerald-50 px-3 py-1 text-[11px] font-semibold text-emerald-700 dark:border-emerald-600 dark:bg-emerald-900/30 dark:text-emerald-300">
-                    <CheckIcon className="h-4 w-4" />
-                    <span>保存済み</span>
-                  </span>
-                )}
-                <button
-                  type="button"
-                  onClick={handleCopy}
-                  className="flex items-center gap-1 rounded-full border border-slate-300 px-3 py-1 text-[11px] font-semibold text-slate-700 transition hover:bg-white dark:border-slate-600 dark:text-slate-100 dark:hover:bg-slate-700"
-                >
-                  {copied ? <CheckIcon className="h-4 w-4" /> : <ClipboardDocumentIcon className="h-4 w-4" />}
-                  <span>{copied ? "コピー済み" : "コピー"}</span>
-                </button>
-              </div>
-            </div>
-            <div className="whitespace-pre-line text-sm leading-6 text-slate-800 dark:text-slate-100">
-              {response.summary ? sanitizeMarkdown(response.summary) : "回答がまだありません。"}
-            </div>
-            {response.bulletPoints?.length ? (
-              <ul className="list-disc space-y-1 pl-4 text-sm text-slate-800 dark:text-slate-100">
-                {response.bulletPoints.map((b, idx) => (
-                  <li key={`${idx}-${b}`}>{sanitizeMarkdown(b)}</li>
-                ))}
-              </ul>
-            ) : null}
-          </div>
+          <AiPanelResponse
+            response={response}
+            saveUrl={saveUrl}
+            saveId={saveId}
+            saved={saved}
+            saving={saving}
+            copied={copied}
+            onSave={handleSave}
+            onCopy={handleCopy}
+          />
         ) : null}
       </div>
     </>
