+++ app/_components/ai-panel/types.ts
@@ -1,0 +1,6 @@
+export type AiResponse = {
+  summary?: string;
+  bulletPoints?: string[];
+  provider?: string;
+  error?: string;
+};
