+++ app/aptitude/_components/aptitude-form.tsx
@@ -5,21 +5,16 @@ import Link from "next/link";
 
 import { AiPanel } from "@/app/_components/ai-panel";
 import { ROUTES } from "@/lib/constants/routes";
-
-type Answers = {
-  interests: string[];
-  strengths: string[];
-  values: string[];
-  enjoy: string;
-  achievements: string;
-  dislike: string;
-  workStyle: string;
-  location: string;
-  industryWish: string;
-  roleWish: string;
-  otherNotes: string;
-  mbti: string;
-};
+import { Answers } from "../types";
+import { CheckboxGroup, SelectBox, TextArea } from "./aptitude-fields";
+import {
+  defaultAnswers,
+  interestOptions,
+  mbtiOptions,
+  strengthOptions,
+  valueOptions,
+} from "./aptitude-constants";
+import { buildPrompt } from "./aptitude-utils";
 
 type Props = {
   initialAnswers: Record<string, unknown> | null;
@@ -27,97 +22,6 @@ type Props = {
   initialResultId: string | null;
 };
 
-const interestOptions = [
-  "SaaS/プロダクト開発",
-  "コンサルティング",
-  "金融/Fintech",
-  "人事/HR",
-  "マーケ/PR",
-  "メディア/コンテンツ",
-  "ヘルスケア/医療",
-  "ゲーム/エンタメ",
-  "公共/教育",
-  "製造/モノづくり",
-  "物流/EC",
-  "スタートアップ/新規事業",
-  "グローバル/海外",
-];
-
-const strengthOptions = [
-  "論理思考・問題解決",
-  "チーム牽引",
-  "コミュニケーション/折衝",
-  "リーダーシップ/推進力",
-  "探究心・学習意欲",
-  "プロジェクトマネジメント",
-  "クリエイティブ/企画",
-  "技術・プログラミング",
-  "営業力・交渉力",
-  "語学/異文化対応",
-];
-
-const valueOptions = [
-  "裁量・意思決定",
-  "安定性",
-  "社会貢献/インパクト",
-  "報酬",
-  "成長スピード",
-  "ワークライフバランス",
-  "リモート/柔軟性",
-  "チームワーク",
-  "専門性の深化",
-  "グローバル環境",
-];
-
-const mbtiOptions = [
-  "ISTJ",
-  "ISFJ",
-  "INFJ",
-  "INTJ",
-  "ISTP",
-  "ISFP",
-  "INFP",
-  "INTP",
-  "ESTP",
-  "ESFP",
-  "ENFP",
-  "ENTP",
-  "ESTJ",
-  "ESFJ",
-  "ENFJ",
-  "ENTJ",
-];
-
-const defaultAnswers: Answers = {
-  interests: [],
-  strengths: [],
-  values: [],
-  enjoy: "",
-  achievements: "",
-  dislike: "",
-  workStyle: "",
-  location: "",
-  industryWish: "",
-  roleWish: "",
-  otherNotes: "",
-  mbti: "",
-};
-
-const buildPrompt = (a: Answers) =>
-  [
-    `興味のある領域: ${a.interests.join(", ") || "未選択"}`,
-    `強み: ${a.strengths.join(", ") || "未選択"}`,
-    `価値観: ${a.values.join(", ") || "未選択"}`,
-    `好きな業務/没頭できること: ${a.enjoy || "未記入"}`,
-    `誇りに思う達成: ${a.achievements || "未記入"}`,
-    `苦手・避けたいこと: ${a.dislike || "未記入"}`,
-    `働き方の希望: ${a.workStyle || "未記入"}`,
-    `希望勤務地/働き方: ${a.location || "未記入"}`,
-    `興味のある業界: ${a.industryWish || "未記入"}`,
-    `興味のある職種: ${a.roleWish || "未記入"}`,
-    `MBTI: ${a.mbti || "未記入"}`,
-    `補足メモ: ${a.otherNotes || "未記入"}`,
-  ].join("\n");
 
 export default function AptitudeForm({ initialAnswers, initialSummary, initialResultId }: Props) {
   const [answers, setAnswers] = useState<Answers>(() => {
@@ -237,93 +141,5 @@ export default function AptitudeForm({ initialAnswers, initialSummary, initialRe
         />
       </div>
     </div>
-  );
-}
-
-function CheckboxGroup({
-  label,
-  options,
-  selected,
-  onToggle,
-}: {
-  label: string;
-  options: string[];
-  selected: string[];
-  onToggle: (value: string) => void;
-}) {
-  return (
-    <div className="space-y-2">
-      <p className="text-sm font-semibold text-slate-800 dark:text-slate-200">{label}</p>
-      <div className="grid grid-cols-1 gap-2 sm:grid-cols-2">
-        {options.map((opt) => (
-          <label
-            key={opt}
-            className="flex cursor-pointer items-center gap-2 rounded-xl border border-slate-200 bg-white px-3 py-2 text-sm text-slate-800 shadow-sm transition hover:border-amber-200 hover:bg-amber-50 dark:border-slate-700 dark:bg-slate-800 dark:text-slate-100"
-          >
-            <input
-              type="checkbox"
-              checked={selected.includes(opt)}
-              onChange={() => onToggle(opt)}
-              className="h-4 w-4 rounded border-slate-300 text-amber-500 focus:ring-amber-400"
-            />
-            <span>{opt}</span>
-          </label>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function TextArea({
-  label,
-  value,
-  onChange,
-}: {
-  label: string;
-  value: string;
-  onChange: (v: string) => void;
-}) {
-  return (
-    <label className="block space-y-2">
-      <span className="text-sm font-medium text-slate-700 dark:text-slate-200">{label}</span>
-      <textarea
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-        rows={3}
-        className="w-full rounded-xl border border-slate-200 bg-white px-4 py-3 text-sm text-slate-900 outline-none focus:border-amber-300 dark:border-slate-700 dark:bg-slate-800 dark:text-slate-100"
-      />
-    </label>
-  );
-}
-
-function SelectBox({
-  label,
-  value,
-  onChange,
-  options,
-  placeholder,
-}: {
-  label: string;
-  value: string;
-  onChange: (v: string) => void;
-  options: string[];
-  placeholder?: string;
-}) {
-  return (
-    <label className="block space-y-2">
-      <span className="text-sm font-medium text-slate-700 dark:text-slate-200">{label}</span>
-      <select
-        value={value}
-        onChange={(e) => onChange(e.target.value)}
-        className="w-full rounded-xl border border-slate-200 bg-white px-4 py-3 text-sm text-slate-900 outline-none focus:border-amber-300 dark:border-slate-700 dark:bg-slate-800 dark:text-slate-100"
-      >
-        <option value="">{placeholder || "選択してください"}</option>
-        {options.map((opt) => (
-          <option key={opt} value={opt}>
-            {opt}
-          </option>
-        ))}
-      </select>
-    </label>
   );
 }
